% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D1.R
\docType{data}
\name{D1}
\alias{D1}
\title{Tablica D1a Konsolidirana bilanca DMFI (aktiva)}
\format{An object of class \code{data.frame} with 295 rows and 29 columns.}
\usage{
D1
}
\description{
Tablica D1a Konsolidirana bilanca DMFI (aktiva)
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
