% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D10.R
\docType{data}
\name{D10}
\alias{D10}
\title{Tablica D10 Inozmena pasiva DMFI}
\format{An object of class \code{data.frame} with 295 rows and 35 columns.}
\usage{
D10
}
\description{
Tablica D10 Inozmena pasiva DMFI
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
