% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D6.R
\docType{data}
\name{D6}
\alias{D6}
\title{Tablica D6 Depozitni novac kod DMFI}
\format{An object of class \code{data.frame} with 295 rows and 12 columns.}
\usage{
D6
}
\description{
Tablica D6 Depozitni novac kod DMFI
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
