% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D8.R
\docType{data}
\name{D8}
\alias{D8}
\title{Tablica D8 Devizni depoziti kod DMFI}
\format{An object of class \code{data.frame} with 295 rows and 31 columns.}
\usage{
D8
}
\description{
Tablica D8 Devizni depoziti kod DMFI
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
