% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/E1.R
\docType{data}
\name{E1}
\alias{E1}
\title{Tablica E1a Agregirana bilanca investicijskih fondova (stanja)}
\format{An object of class \code{data.frame} with 83 rows and 36 columns.}
\usage{
E1
}
\description{
Tablica E1a Agregirana bilanca investicijskih fondova (stanja)
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
