% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1.R
\docType{data}
\name{F1}
\alias{F1}
\title{Tablica F1 Aktivne kamatne stope HNB-a}
\format{An object of class \code{data.frame} with 313 rows and 11 columns.}
\usage{
F1
}
\description{
Tablica F1 Aktivne kamatne stope HNB-a
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
