% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F3.R
\docType{data}
\name{F3}
\alias{F3}
\title{Tablica F3 Obveze pricuve banaka}
\format{An object of class \code{data.frame} with 311 rows and 10 columns.}
\usage{
F3
}
\description{
Tablica F3 Obveze pricuve banaka
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
