% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G10a.R
\docType{data}
\name{G10a}
\alias{G10a}
\alias{G10aG}
\title{Tablica G10a Godisnji prosjeci srednjih deviznih tecajeva}
\format{An object of class \code{data.frame} with 26 rows and 9 columns.}
\usage{
G10aG
}
\description{
Tablica G10a Godisnji prosjeci srednjih deviznih tecajeva
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
