% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G10a.R
\docType{data}
\name{G10aM}
\alias{G10aM}
\title{Tablica G10a Mjesecni prosjeci srednjih deviznih tecajeva}
\format{An object of class \code{data.frame} with 324 rows and 9 columns.}
\usage{
G10aM
}
\description{
Tablica G10a Mjesecni prosjeci srednjih deviznih tecajeva
}
\keyword{datasets}
