% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G10b.R
\docType{data}
\name{G10b}
\alias{G10b}
\alias{G10bG}
\title{Tablica G10b Godisnji srednji devizni tecajevi HNB-a na kraju razdoblja}
\format{An object of class \code{data.frame} with 26 rows and 11 columns.}
\usage{
G10bG
}
\description{
Tablica G10b Godisnji srednji devizni tecajevi HNB-a na kraju razdoblja
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
