% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G10b.R
\docType{data}
\name{G10bM}
\alias{G10bM}
\title{Tablica G10b Mjesecni srednji devizni tecajevi HNB-a na kraju razdoblja}
\format{An object of class \code{data.frame} with 324 rows and 9 columns.}
\usage{
G10bM
}
\description{
Tablica G10b Mjesecni srednji devizni tecajevi HNB-a na kraju razdoblja
}
\keyword{datasets}
