% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G8a.R
\docType{data}
\name{G8aG}
\alias{G8aG}
\title{Tablica G8aG Kamatne stope na trezorske zapise RH (godisnji)}
\format{An object of class \code{data.frame} with 23 rows and 6 columns.}
\usage{
G8aG
}
\description{
Tablica G8aG Kamatne stope na trezorske zapise RH (godisnji)
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
