% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G8a.R
\docType{data}
\name{G8aM}
\alias{G8aM}
\title{Tablica G8aM Kamatne stope na trezorske zapise RH (mjesecni)}
\format{An object of class \code{data.frame} with 109 rows and 6 columns.}
\usage{
G8aM
}
\description{
Tablica G8aM Kamatne stope na trezorske zapise RH (mjesecni)
}
\keyword{datasets}
