% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G8b.R
\docType{data}
\name{G8bG}
\alias{G8bG}
\title{Tablica G8aG Prinosi do dospijeca na obveznice (godisnji)}
\format{An object of class \code{data.frame} with 18 rows and 45 columns.}
\usage{
G8bG
}
\description{
Tablica G8aG Prinosi do dospijeca na obveznice (godisnji)
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
