% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G8b.R
\docType{data}
\name{G8bM}
\alias{G8bM}
\title{Tablica G8aM Prinosi do dospijeca na obveznice (mjesecni)}
\format{An object of class \code{data.frame} with 209 rows and 45 columns.}
\usage{
G8bM
}
\description{
Tablica G8aM Prinosi do dospijeca na obveznice (mjesecni)
}
\keyword{datasets}
