% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H1.R
\docType{data}
\name{H1}
\alias{H1}
\alias{H1a}
\title{Tablica H1 Platna bilanca - svodna tablica (EUR)}
\format{An object of class \code{data.frame} with 93 rows and 34 columns.}
\usage{
H1a
}
\description{
Tablica H1 Platna bilanca - svodna tablica (EUR)
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
