% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H11.R
\docType{data}
\name{h11}
\alias{h11}
\alias{H11}
\title{Tablica H11 Indeksi efektivnih tecajeva kune}
\format{An object of class \code{data.frame} with 276 rows and 6 columns.}
\usage{
H11
}
\description{
Tablica H11 Indeksi efektivnih tecajeva kune
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
