% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H1.R
\docType{data}
\name{H1b}
\alias{H1b}
\title{Tablica H1 Platna bilanca - svodna tablica (USD)}
\format{An object of class \code{data.frame} with 93 rows and 34 columns.}
\usage{
H1b
}
\description{
Tablica H1 Platna bilanca - svodna tablica (USD)
}
\keyword{datasets}
