% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H6.R
\docType{data}
\name{H6}
\alias{H6}
\title{Tablica H6 Platna bilanca - svodna tablica (HRK)}
\format{An object of class \code{data.frame} with 93 rows and 34 columns.}
\usage{
H6
}
\description{
Tablica H6 Platna bilanca - svodna tablica (HRK)
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
