% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H7.R
\docType{data}
\name{H7}
\alias{H7}
\title{Tablica H7 Medunarodne pricuve HNB-a i devizne pricuve banaka}
\format{An object of class \code{data.frame} with 313 rows and 9 columns.}
\usage{
H7
}
\description{
MPHNB = Medunarodne pricuve Hrvatske narodne banke
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
