% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H8.R
\docType{data}
\name{H8a}
\alias{H8a}
\title{Tablica H8a Medunarodne pricuve i devizna likvidnost
Sluzbene medunarodne pricuve i ostale devizne pricuve (priblizna trzisna vrijednost)}
\format{An object of class \code{data.frame} with 228 rows and 19 columns.}
\usage{
H8a
}
\description{
Tablica H8a Medunarodne pricuve i devizna likvidnost
Sluzbene medunarodne pricuve i ostale devizne pricuve (priblizna trzisna vrijednost)
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
