% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H8.R
\docType{data}
\name{H8b}
\alias{H8b}
\title{Tablica H8b Medunarodne pricuve i devizna likvidnost}
\format{An object of class \code{data.frame} with 228 rows and 34 columns.}
\usage{
H8b
}
\description{
Obvezani kratkorocni neto odljevi medunarodnih i ostalih deviznih pricuva (nominalna vrijednost)
}
\keyword{datasets}
