% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H8.R
\docType{data}
\name{H8c}
\alias{H8c}
\title{Tablica H8c Monetarni i kreditni agregati}
\format{An object of class \code{data.frame} with 228 rows and 20 columns.}
\usage{
H8c
}
\description{
Potencijalni kratkorocni neto odljevi medunarodnih pricuva (nominalna vrijednost)
}
\keyword{datasets}
