% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I1.R
\docType{data}
\name{I1}
\alias{I1}
\title{Tablica I1a Nefinancijski racuni opce drzave}
\format{An object of class \code{data.frame} with 90 rows and 83 columns.}
\usage{
I1
}
\description{
Tablica I1a Nefinancijski racuni opce drzave
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
