% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I1.R
\docType{data}
\name{I1b}
\alias{I1b}
\title{Tablica I1b Nefinancijski racuni opce drzave}
\format{An object of class \code{data.frame} with 63 rows and 88 columns.}
\usage{
I1b
}
\description{
Tablica I1b Nefinancijski racuni opce drzave
}
\keyword{datasets}
