% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I2.R
\docType{data}
\name{I2}
\alias{I2}
\title{Tablica I2 Glavni makroagregati opce drzave u skupu racuna ESA 2010}
\format{An object of class \code{data.frame} with 16 rows and 44 columns.}
\usage{
I2
}
\description{
Tablica I2 Glavni makroagregati opce drzave u skupu racuna ESA 2010
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
