% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I3.R
\docType{data}
\name{I3}
\alias{I3}
\title{Tablica I3 Dug opce drzave (ESA 2010)}
\format{An object of class \code{data.frame} with 274 rows and 31 columns.}
\usage{
I3
}
\description{
Tablica I3 Dug opce drzave (ESA 2010)
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
