% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J1.R
\docType{data}
\name{J1}
\alias{J1}
\title{Tablica J1 Indeksi potrosackih cijena i proizvodackih cijena industrije}
\format{An object of class \code{data.frame} with 251 rows and 13 columns.}
\usage{
J1
}
\description{
Tablica J1 Indeksi potrosackih cijena i proizvodackih cijena industrije
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
