% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J1a.R
\docType{data}
\name{J1a}
\alias{J1a}
\title{Tablica J1a Harmonizirani indeksi potrosackih cijena}
\format{An object of class \code{data.frame} with 251 rows and 19 columns.}
\usage{
J1a
}
\description{
Tablica J1a Harmonizirani indeksi potrosackih cijena
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
