% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J2.R
\docType{data}
\name{J2}
\alias{J2}
\title{Tablica J2 Temeljni indeksi potrosakcih cijena}
\format{An object of class \code{data.frame} with 251 rows and 10 columns.}
\usage{
J2
}
\description{
Tablica J2 Temeljni indeksi potrosakcih cijena
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
