% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J3.R
\docType{data}
\name{J3}
\alias{J3}
\alias{J3g}
\title{Tablica J3m Indeksi cijena stambenih nekretnina (godisnje)}
\format{An object of class \code{data.frame} with 16 rows and 13 columns.}
\usage{
J3g
}
\description{
Tablica J3m Indeksi cijena stambenih nekretnina (godisnje)
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
