% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J4.R
\docType{data}
\name{J4}
\alias{J4}
\title{Tablica J4 Prosjecne mjesecne neto place}
\format{An object of class \code{data.frame} with 322 rows and 5 columns.}
\usage{
J4
}
\description{
Tablica J4 Prosjecne mjesecne neto place
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
