% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/J5.R
\docType{data}
\name{J5}
\alias{J5}
\title{Tablica J5 Indeksi pouzdanja ocekivanja i raspolozenja potrosaca}
\format{An object of class \code{data.frame} with 188 rows and 11 columns.}
\usage{
J5
}
\description{
Tablica J5 Indeksi pouzdanja ocekivanja i raspolozenja potrosaca
}
\references{
\url{https://www.hnb.hr/statistika}
}
\author{
David Ravnik \email{dravnik@unipu.hr}

Tomislava Kolar \email{tkolar@unipu.hr}
}
\keyword{dataframe}
