% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizualizacijaA1.R
\name{plotA1}
\alias{plotA1}
\title{Plotanje A1}
\usage{
plotA1(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot a1
}
\description{
Funkcija A1
}
