% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizualizacijaBDP.R
\name{plotBDPgodina}
\alias{plotBDPgodina}
\title{Plotanje godisnjeg rasta BDP-a}
\usage{
plotBDPgodina(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot godisnjeg BPD-a
}
\description{
Funkcija koja plota godisnji rast BDP-a
}
