% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizualizacijaBDP.R
\name{plotBDPkvartal1}
\alias{plotBDPkvartal1}
\title{Plotanje kvartalni rasta BDP-a}
\usage{
plotBDPkvartal1(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot kvartalnog BPD-a
}
\description{
Funkcija koja plota kvartalni rast BDP-a
}
