% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizualizacijaC1AKTIVA-PASIVA.R
\name{plotBilancaaktiva}
\alias{plotBilancaaktiva}
\title{Plotanje c1}
\usage{
plotBilancaaktiva(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot c1
}
\description{
Funkcija c1
}
