% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizualizacijaD10.R
\name{plotD10Inozmenapasiva}
\alias{plotD10Inozmenapasiva}
\title{Plotanje D10}
\usage{
plotD10Inozmenapasiva(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot D10
}
\description{
Funkcija koja plota D10
}
