% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizualizacijaD6.R
\name{plotDepozitninovacD6}
\alias{plotDepozitninovacD6}
\title{Plotanje d6}
\usage{
plotDepozitninovacD6(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot d6
}
\description{
Funkcija d6
}
