% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizualizacijaD8dnevnidepozit.R
\name{plotDnevniDepozitiD8}
\alias{plotDnevniDepozitiD8}
\title{Plotanje d8}
\usage{
plotDnevniDepozitiD8(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot d8
}
\description{
Funkcija d8
}
