% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizualizacijaE1AgregiranaBilanca.R
\name{plotE1Agregiranabilanca}
\alias{plotE1Agregiranabilanca}
\title{Plotanje e1}
\usage{
plotE1Agregiranabilanca(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot e1
}
\description{
Funkcija koja plota e1
}
