% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizualizacijaF1.R
\name{plotF1}
\alias{plotF1}
\title{Plotanje f1}
\usage{
plotF1(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot f1
}
\description{
Funkcija f1
}
