% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizualizacijaF3.R
\name{plotF3}
\alias{plotF3}
\title{Plotanje f3}
\usage{
plotF3(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot f3
}
\description{
Funkcija koja plota f3
}
