% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vizualizacijal1.R
\name{plotIlaNefinancijskiracuni}
\alias{plotIlaNefinancijskiracuni}
\title{Plotanje l1}
\usage{
plotIlaNefinancijskiracuni(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot l1
}
\description{
Funkcija koja plota l1
}
