% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizualizacijaD2InozemnaAktiva.R
\name{plotInozemnaAktiva1}
\alias{plotInozemnaAktiva1}
\title{Plotanje d2}
\usage{
plotInozemnaAktiva1(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot d2
}
\description{
Funkcija d2
}
