% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizualizacijaJ1.R
\name{plotJ1}
\alias{plotJ1}
\title{Plotanje j1}
\usage{
plotJ1(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot j1
}
\description{
Funkcija koja plota j1
}
