% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizualizacijaJ2.R
\name{plotJ2}
\alias{plotJ2}
\title{Plotanje j2}
\usage{
plotJ2(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot j2
}
\description{
Funkcija koja plota j2
}
