% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizualizacijaJ3.R
\name{plotJ3}
\alias{plotJ3}
\title{Plotanje j3}
\usage{
plotJ3(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot j3
}
\description{
Funkcija koja plota j3
}
