% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizuačizacijaJ5.R
\name{plotJ5}
\alias{plotJ5}
\title{Plotanje j5}
\usage{
plotJ5(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot j5
}
\description{
Funkcija j5
}
