% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizualizacijaD1iD1a.R
\name{plotKonsolidiranabilancaD1}
\alias{plotKonsolidiranabilancaD1}
\title{Plotanje gd1}
\usage{
plotKonsolidiranabilancaD1(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot d1
}
\description{
Funkcija d1
}
