% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizualizacijaKunskiDepozitd7.R
\name{plotKunskiDepozit1}
\alias{plotKunskiDepozit1}
\title{Plotanje d7}
\usage{
plotKunskiDepozit1(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot d7
}
\description{
Funkcija koja plota d7
}
