% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizualizacijaPasivneKamatneStopef2.R
\name{plotPasivneKamatneStope}
\alias{plotPasivneKamatneStope}
\title{Plotanje f2}
\usage{
plotPasivneKamatneStope(stupac, Naziv)
}
\arguments{
\item{stupac}{predstavlja stupac kojeg korisnik zeli vizualizirati}

\item{nazivStupca}{predstavlja naziv stupac kojeg korisnik zeli vizualizirati}
}
\value{
plot f2
}
\description{
Funkcija koja plota f2
}
